Scrape decisions from the Trial Decision Library.
This scrapes the table at https://nypdonline.org/link/1016 . Most of these should already be included in the profile data, but in the case a trial decision gets posted when an officer no longer works for the department (e.g. they got fired) it'll show up here but not in the profile data (because their profile goes away when they leave).