Also sort trial decision documents by officer name.
This will remove some order non-determinism between two records with the same date and URL, but different officers.

This prioritizes the officer names over the URL so things stay in the same order even if the URL switches back and forth (I think we saw this happening?).

Note that this will result in a one-time reordering, as records on the same date will now first be ordered by officer, then by URL. After that, I'm hoping there won't be any more non-determinism.